requirements.txt
argon2-cffi==19.2.0
astroid==2.3.3
bcrypt==3.1.7
certifi==2018.8.24
cffi==1.14.0
colorama==0.4.3
dj-database-url==0.5.0
Django==2.2.10
django-heroku==0.3.1
django-nucleus==0.1.2
djangorestframework==3.11.0
Faker==4.0.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
psycopg2==2.8.5
pycparser==2.19
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.1
whitenoise==5.1.0
wincertstore==0.2
wrapt==1.11.2
